Hades_Lite/Hades Web/hades_core/requirements.txt
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dateutil>=2.8.0
